requirements.txt
pandas
pillow
tqdm
tensorflow
keras
sklearn
matplotlib